backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings


langchain
langchain-openai
langchain-milvus
langchain-community
langchain-core

pymilvus

openai

python-multipart
aiohttp
tiktoken
numpy